sprint2-liu-modified exception logic and new api for get all locations and colleges

+++ NUboard-services/src/main/java/com/neu/nuboard/model/College.java
@@ -1,6 +1,7 @@
 package com.neu.nuboard.model;
 
 import jakarta.persistence.*;
+import org.springframework.web.bind.annotation.CrossOrigin;
 
 import java.util.List;
 
@@ -10,7 +11,7 @@ import java.util.List;
 public class College {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(unique = true, nullable = false)
     private String name;
@@ -21,7 +22,7 @@ public class College {
     private List<User> users;
 
     //Getter
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
@@ -34,7 +35,7 @@ public class College {
     }
 
     // Setter
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
